1) Refactoring of game (from single file to multiple files with folders) 2) Added header 3) Added routing
