+++ PetFamily.Frontend/src/app/RootLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router";
+import { Header } from "../components/Header";
+
+export function RootLayout() {
+	return (
+		<div>
+			<Header />
+			<Outlet />
+		</div>
+	);
+}
