+++ PetFamily.Frontend/src/index.css
@@ -1,59 +1,9 @@
-* {
-	box-sizing: border-box;
-}
+@import "tailwindcss";
 
-body {
-	font-family: sans-serif;
-	margin: 20px;
-	padding: 0;
-}
-
-h1 {
-	margin-top: 0;
-	font-size: 22px;
-}
-
-h2 {
-	margin-top: 0;
-	font-size: 20px;
-}
-
-h3 {
-	margin-top: 0;
-	font-size: 18px;
-}
-
-h4 {
-	margin-top: 0;
-	font-size: 16px;
-}
-
-h5 {
-	margin-top: 0;
-	font-size: 14px;
-}
-
-h6 {
-	margin-top: 0;
-	font-size: 12px;
-}
-
-code {
-	font-size: 1.2em;
-}
-
-ul {
-	padding-inline-start: 20px;
-}
-
-* {
-	box-sizing: border-box;
-}
-
-body {
-	font-family: sans-serif;
-	margin: 20px;
-	padding: 0;
+@layer utilities {
+	.dark {
+		color-scheme: dark;
+	}
 }
 
 .square {
