+++ PetFamily.Frontend/src/pages/Game/components/Game.tsx
@@ -1,0 +1,40 @@
+import { useState } from "react";
+import { BoardArray } from "../../types/gameTypes";
+import { Board } from "./Board";
+
+export default function Game() {
+	const [history, setHistory] = useState<BoardArray[]>([Array(9).fill(null)]);
+	const [currentMove, setCurrentMove] = useState(0);
+	const currentSquares = history[currentMove];
+	const xIsNext = currentMove % 2 === 0;
+
+	const handlePlay = (nextSquares: BoardArray) => {
+		const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
+		setHistory(nextHistory);
+		setCurrentMove(nextHistory.length - 1);
+	};
+
+	const jumpTo = (nextMove: number) => {
+		setCurrentMove(nextMove);
+	};
+
+	const moves = history.map((_, move) => {
+		const description = move > 0 ? `Go to move #${move}` : "Go to game start";
+		return (
+			<li key={move}>
+				<button onClick={() => jumpTo(move)}>{description}</button>
+			</li>
+		);
+	});
+
+	return (
+		<div className="game">
+			<div className="game-board">
+				<Board xIsNext={xIsNext} squares={currentSquares} onPlay={handlePlay} />
+			</div>
+			<div className="game-info">
+				<ol>{moves}</ol>
+			</div>
+		</div>
+	);
+}
