+++ PetFamily.Frontend/src/pages/Game/components/Square.tsx
@@ -1,0 +1,10 @@
+import { Button } from "@mui/material";
+import { SquareProps } from "../../types/gameTypes";
+
+export function Square({ value, onSquareClick }: SquareProps) {
+	return (
+		<Button variant="outlined" onClick={onSquareClick}>
+			{value || "\u00A0"}
+		</Button>
+	);
+}
