+++ PetFamily.Frontend/src/pages/utils/calculateWinner.tsx
@@ -1,0 +1,21 @@
+import { BoardArray, SquareValue } from "./Game";
+
+export function calculateWinner(squares: BoardArray): SquareValue {
+	const lines = [
+		[0, 1, 2],
+		[3, 4, 5],
+		[6, 7, 8],
+		[0, 3, 6],
+		[1, 4, 7],
+		[2, 5, 8],
+		[0, 4, 8],
+		[2, 4, 6],
+	];
+
+	for (const [a, b, c] of lines) {
+		if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+			return squares[a];
+		}
+	}
+	return null;
+}
