+++ PetFamily.Frontend/src/theme.ts
@@ -1,0 +1,14 @@
+import { cyan } from "@mui/material/colors";
+import { createTheme } from "@mui/material/styles";
+
+export const darkTheme = createTheme({
+	palette: {
+		mode: "dark",
+		primary: {
+			light: cyan[400],
+			main: cyan[500],
+			dark: cyan[600],
+			contrastText: "#fff",
+		},
+	},
+});
